userService and CreditService interfaces added
