+++ src/main/java/com/example/demo/service/UserService.java
@@ -15,40 +15,19 @@ import java.util.Set;
 /**
  * Created by İbrahim Başar YARGICI at 28.09.2021
  */
-@Service
-public class UserService implements UserDetailsService {
-    private final UserRepository userRepository;
 
-    public UserService(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
+public interface UserService {
 
-    public Set<User> getAll() {
-        return new HashSet<>(userRepository.findAll());
-    }
 
-    public User findUserById(int id) {
-        // TODO business code
-        return userRepository.getById(id);
-    }
+     Set<User> getAll();
 
-    public User save(User user) {
-        // TODO business code
-       return this.userRepository.save(user);
-    }
+     User findUserById(int id);
 
-    public User update(User user) {
-        // TODO business code
-        return userRepository.save(user);
-    }
+    User save(User user);
 
-    public void delete(int id) {
-        // TODO business code
-        userRepository.delete(findUserById(id));
-    }
+     User update(User user);
 
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        return null;
-    }
+     void delete(int id);
+
+
 }
