+++ src/main/java/com/example/demo/service/impl/UserServiceImpl.java
@@ -1,4 +1,51 @@
 package com.example.demo.service.impl;
 
-public class UserServiceImpl {
+import com.example.demo.model.User;
+import com.example.demo.repository.UserRepository;
+import com.example.demo.service.UserService;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+import java.util.HashSet;
+import java.util.Set;
+
+@Service
+public class UserServiceImpl implements UserService, UserDetailsService {
+    private final UserRepository userRepository;
+
+    public UserServiceImpl(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
+
+    @Override
+    public Set<User> getAll() {
+        return new HashSet<>(userRepository.findAll());
+    }
+
+    @Override
+    public User findUserById(int id) {
+    return this.userRepository.getById(id);
+    }
+
+    @Override
+    public User save(User user) {
+        return this.userRepository.save(user);
+    }
+
+    @Override
+    public User update(User user) {
+        return this.userRepository.save(user);
+    }
+
+    @Override
+    public void delete(int id) {
+        this.userRepository.deleteById(id);
+    }
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        return null;
+    }
 }
